Sign Up Validation and Error Fixing
