+++ login.php
@@ -6,7 +6,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="shortcut icon" href="./assets/pics and icons/favicon.png">
     <link rel="stylesheet" href="./assets/reg.css">
-    <title>Document</title>
+    <title>Unigram | Log In</title>
 </head>
 
 <body>
